Don't wait the activity to be completely launched.
Appium APK test waits the launchable activity to be focused, but some apps switches to another activity immediately after the init activity is called and the wait never ends and then test fails.
On the other hand, Appium package/activity test never waits the launchable activity to be focused, and the wait problem does not happen. And it seems that this does not cause any critical issue.
So, to avoid APK wait problem, we always skip the wait process to specify "appWaitActivity": "*"